requirements.txt
fastapi~=0.112.2
torch~=2.5.0
pillow~=10.2.0
ultralytics~=8.3.23
uvicorn~=0.23.2
python-multipart~=0.0.20